refactor: moving Bot into the common package
