+++ azurerm/internal/clients/client_bot.go
@@ -1,17 +1,17 @@
-package bot
+package clients
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/preview/botservice/mgmt/2018-07-12/botservice"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/common"
 )
 
-type Client struct {
+type BotClient struct {
 	BotClient        *botservice.BotsClient
 	ConnectionClient *botservice.BotConnectionClient
 	ChannelClient    *botservice.ChannelsClient
 }
 
-func BuildClient(o *common.ClientOptions) *Client {
+func newBotClient(o *common.ClientOptions) *BotClient {
 	botClient := botservice.NewBotsClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&botClient.Client, o.ResourceManagerAuthorizer)
 
@@ -21,7 +21,7 @@ func BuildClient(o *common.ClientOptions) *Client {
 	channelClient := botservice.NewChannelsClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&channelClient.Client, o.ResourceManagerAuthorizer)
 
-	return &Client{
+	return &BotClient{
 		BotClient:        &botClient,
 		ChannelClient:    &channelClient,
 		ConnectionClient: &connectionClient,
